Move lambda tests to lambdas/
